refactor: add getGatewayAPITracfficRoutingConfig function
