+++ internal/defaults/defaults.go
@@ -1,0 +1,3 @@
+package defaults
+
+const ConfigMap = "argo-gatewayapi-configmap"
