+++ internal/utils/common.go
